% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{Bhat_mat_rlist}
\alias{Bhat_mat_rlist}
\title{Generate a list of rank-specific Bhat matrices (the coefficient of Ridge Redundancy Analysis for each parameter lambda and nrank).}
\usage{
Bhat_mat_rlist(Bhat_comp_1L, nrank)
}
\arguments{
\item{Bhat_comp_1L}{A list containing components of Bhat for each lambda value.}

\item{nrank}{A numeric vector indicating the rank(s) of Bhat.}
}
\value{
A list of matrices, each representing a rank-specific Bhat matrix.
}
\description{
Generate a list of rank-specific Bhat matrices (the coefficient of Ridge Redundancy Analysis for each parameter lambda and nrank). In our formula, Bhat is stored as LeftBhatlambda_k (p times r matrix) and RightBhatlambda_k (q times r). Here, n is the number of samples, p is the number of variables of X, q is the number of variables of Y, and r is the specified rank in the Ridge Redundancy Analysis.
}
