% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{get_lambda}
\alias{get_lambda}
\title{Estimate an appropriate value for the ridge penalty (lambda).}
\usage{
get_lambda(Y, X, scale.X = FALSE, sample.X = sample.X, sample.Y = sample.Y)
}
\arguments{
\item{Y}{A numeric matrix of response variables.}

\item{X}{A numeric matrix of explanatory variables.}

\item{scale.X}{Logical indicating if \code{X} is scaled. Default is \code{FALSE}.}

\item{sample.X}{A number of variables sampled from X for the lamdba range estimate. Default is 1000.}

\item{sample.Y}{A number of variables sampled from Y for the lamdba range estimate. Default is 1000.}
}
\value{
A numeric vector of the range of lambda values.
}
\description{
Estimate an appropriate value for the ridge penalty (lambda).
}
