% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{get_rlist}
\alias{get_rlist}
\title{Generate rank-specific matrices by combining the left and right components.}
\usage{
get_rlist(LV, RV, nrank)
}
\arguments{
\item{LV}{A numeric matrix of the left component vectors.}

\item{RV}{A numeric matrix of the right component vectors.}

\item{nrank}{A numeric vector indicating the rank(s) of Bhat.}
}
\value{
A list of matrices of rank-specific combinations of LV and RV.
}
\description{
Generate rank-specific matrices by combining the left and right components of the coefficeint Bhat. In our formula, Bhat is stored as LeftBhatlambda_k (the left component vectors, p times r matrix) and RightBhatlambda_k (the right component vectors, q times r) for each lambda value. Here, n is the number of samples, p is the number of variables of X, q is the number of variables of Y, and r is the specified rank in the Ridge Redundancy Analysis.
}
