% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{rrda.summary}
\alias{rrda.summary}
\title{Summarize the results of cross-validation for the coefficient Bhat obtained from the \code{rrda.cv} function.}
\usage{
rrda.summary(cv_result = cv_result)
}
\arguments{
\item{cv_result}{A result list from the function \code{rrda.cv}, containing a matrix of MSE values for each rank and lambda, and a vector of lambda values.}
}
\value{
Prints the estimated Coefficients, rank, lambda, and MSE from cross-validation.
}
\description{
The function provides a summary of the results from cross-validation for Bhat obtained using the rrda.cv function. It displays the Mean Squared Error (MSE), rank, and lambda values for different cross-validation options.
}
