% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{sqrt_inv_d2_lambda}
\alias{sqrt_inv_d2_lambda}
\title{Compute the square root of the inverse of (d^2 + lambda).}
\usage{
sqrt_inv_d2_lambda(d, lambda)
}
\arguments{
\item{d}{A scalar of singular value of X.}

\item{lambda}{A scalar of the ridge penalty.}
}
\value{
A scalar of the square root of the inverse of (d^2 + lambda).
}
\description{
Compute the square root of the inverse of (d^2 + lambda).
}
