% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{unbiased_scale}
\alias{unbiased_scale}
\title{Scale a matrix using unbiased estimators for the mean and standard deviation.}
\usage{
unbiased_scale(x)
}
\arguments{
\item{x}{A numeric matrix to be scaled.}
}
\value{
A scaled numeric matrix.
}
\description{
Scale a matrix using unbiased estimators for the mean and standard deviation.
}
