% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{unscale_matrices}
\alias{unscale_matrices}
\title{Unscale a matrix based on provided mean and standard deviation values.}
\usage{
unscale_matrices(mat, means, std = NULL)
}
\arguments{
\item{mat}{A numeric matrix to be unscaled.}

\item{means}{A numeric vector of means for each column.}

\item{std}{A numeric vector of standard deviations for each column (optional).}
}
\value{
An unscaled numeric matrix.
}
\description{
Unscale a matrix based on provided mean and standard deviation values.
}
