% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{unscale_nested_matrices_map}
\alias{unscale_nested_matrices_map}
\title{Apply unscaling to a nested list of matrices using specified mean and standard deviation values.}
\usage{
unscale_nested_matrices_map(mat, means, std = NULL)
}
\arguments{
\item{mat}{A nested list of matrices.}

\item{means}{A numeric vector of means for each matrix.}

\item{std}{A numeric vector of standard deviations for each matrix (optional).}
}
\value{
A nested list of unscaled matrices.
}
\description{
Apply unscaling to a nested list of matrices using specified mean and standard deviation values.
}
