% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{Yhat_mat_rlist}
\alias{Yhat_mat_rlist}
\title{Generate a list of rank-specific Yhat matrices.}
\usage{
Yhat_mat_rlist(Bhat_comp_1L, X, nrank)
}
\arguments{
\item{Bhat_comp_1L}{A list containing components of Bhat.}

\item{X}{A numeric matrix of explanatory variables.}

\item{nrank}{A numeric vector indicating the rank(s) of Bhat.}
}
\value{
A list of matrices, each representing rank-specific predicted values Yhat.
}
\description{
Generate a list of rank-specific Yhat matrices.
}
